+++ app/src/main/java/com/example/rmp_frontend/domain/model/TradeResult.kt
@@ -3,7 +3,7 @@ package com.example.rmp_frontend.domain.model
 data class TradeResult(
     val orderId: String,
     val instrumentId: String,
-    val quantity: Int,
+    val quantity: Double,
     val executedPrice: Double,
     val type: TransactionType,
     val message: String,
